pyproject.toml
[project]
name = "dtc-ml-zoomcamp-capstone"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.30",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "scikit-learn>=1.8.0",
    "xgboost-cpu>=3.1.3",
]

[dependency-groups]
dev = [
    "seaborn>=0.13.2",
    "jupyterlab>=4.5.2",
]
